server/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "sni"
version = "2.0"
description = "Content server for SNI"
dependencies = [
    "babel==2.12.1",
    "flask==2.3.3",
    "flask-caching==2.0.2",
    "flask-flatpages==0.8.1",
    "flask-sqlalchemy==3.1.1",
    "lxml==4.9.3",
    "pydantic==2.3.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
]

[project.optional-dependencies]
dev = ["black==23.3.0", "ruff==0.0.292"]
prod = ["gunicorn==21.2.0"]

[tool.ruff]
select = ["E", "F", "I"]
